- introduced the new file ff.hpp containing all includes of the building blocks - removed warnings - introduced combine_with_firststage and combine_with_laststage - minor fixes

+++ tests/test_spinBarrier.cpp
@@ -79,7 +79,7 @@ public:
         }
         return EOS;
     }
-    void set_id(int id) {
+    void set_id(ssize_t id) {
         ff_node::set_id(id);
     }
     int run(bool=false)  { return ff_node::run();}
